LayoutTests/imported/w3c/web-platform-tests/xhr/access-control-basic-denied-expected.txt

CONSOLE MESSAGE: line 25: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 25: XMLHttpRequest cannot load http://127.0.0.1:8800/xhr/resources/access-control-basic-denied.py due to access control checks.

PASS Same-origin request accepted 
PASS Cross-origin request denied 
